refactor: move createBlockHub
