+++ packages/blocks/src/page-block/default/utils.ts
@@ -3,14 +3,13 @@ import {
   BLOCK_ID_ATTR,
   BLOCK_SERVICE_LOADING_ATTR,
 } from '@blocksuite/global/config';
-import { assertExists, matchFlavours, Signal } from '@blocksuite/global/utils';
+import { assertExists, matchFlavours } from '@blocksuite/global/utils';
 import type { BaseBlockModel } from '@blocksuite/store';
-import { BlockHub, DragHandle } from '../../components/index.js';
+import { DragHandle } from '../../components/index.js';
 import { toast } from '../../components/toast.js';
 import type { EmbedBlockModel } from '../../embed-block/embed-model.js';
 import { getService } from '../../__internal__/service.js';
 import {
-  asyncFocusRichText,
   doesInSamePath,
   getBlockById,
   getBlockElementByModel,
@@ -21,7 +20,6 @@ import type {
   CodeBlockOption,
   DefaultPageBlockComponent,
 } from './default-page-block.js';
-import type { EdgelessPageBlockComponent } from '../edgeless/edgeless-page-block.js';
 
 export interface EditingState {
   model: BaseBlockModel;
@@ -469,39 +467,4 @@ export function createDragHandle(defaultPageBlock: DefaultPageBlockComponent) {
       );
     },
   });
-}
-
-export function createBlockHub(
-  pageBlock: DefaultPageBlockComponent | EdgelessPageBlockComponent,
-  updateSelectedRects?: Signal<DOMRect[]>
-) {
-  const page = pageBlock.page;
-  return new BlockHub(
-    {
-      onDropCallback: (e, end) => {
-        const dataTransfer = e.dataTransfer;
-        assertExists(dataTransfer);
-        const data = dataTransfer.getData('affine/block-hub');
-        const blockProps = JSON.parse(data);
-        if (blockProps.flavour === 'affine:database') {
-          if (!page.awarenessStore.getFlag('enable_database')) {
-            console.warn('database block is not enabled');
-            return;
-          }
-        }
-        const targetModel = end.model;
-        const rect = end.position;
-        page.captureSync();
-        const distanceToTop = Math.abs(rect.top - e.y);
-        const distanceToBottom = Math.abs(rect.bottom - e.y);
-        const id = page.addSiblingBlock(
-          targetModel,
-          blockProps,
-          distanceToTop < distanceToBottom ? 'right' : 'left'
-        );
-        asyncFocusRichText(page, id);
-      },
-    },
-    updateSelectedRects
-  );
 }
