+++ packages/blocks/src/page-block/utils/index.ts
@@ -2,3 +2,4 @@ export * from './bind-hotkey.js';
 export * from './container-operations.js';
 export { getDragDirection } from './position.js';
 export type { DragDirection } from './position.js';
+export * from './components.js';
